requirements.txt
altgraph==0.17.4
click==8.1.7
cmake==3.31.1
colorama==0.4.6
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
greenlet==3.1.1
importlib_metadata==8.5.0
iniconfig==2.0.0
mysql-connector-python==9.0.0
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
pluggy==1.5.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pytest==8.3.4
pywin32-ctypes==0.2.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
zipp==3.20.2
